search bar and registration, login and logout are done

+++ base/views.py
@@ -1,5 +1,12 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+from .models import Room, Topic
 from .forms import RoomForm
 
 # rooms = [
@@ -9,9 +16,56 @@ from .forms import RoomForm
 #     {'id':4, 'name':"mina emad lolz", },
 # ]
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username = username)
+        except:
+            messages.error(request, 'User does not exist!')
+        user = authenticate(request, username = username, password = password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist!')
+    context = {'page': page}
+    return render(request, 'base/login_registration_page.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registrationPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Error happened while registration')
+
+    context = {'page': page, 'form': form}
+    return render(request, 'base/login_registration_page.html', context)
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms':rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(Q(topic__name__icontains = q) | Q(name__icontains = q) | Q(description__icontains = q))
+    topics = Topic.objects.all()
+    rooms_count = rooms.count()
+    context = {'rooms':rooms, 'topics': topics, 'rooms_count': rooms_count}
     return render(request,'base/home.html' , context)
 
 def room(request,pk):
@@ -24,6 +78,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url = 'login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,9 +91,13 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def update_room(request, pk):
     room = Room.objects.get(id=int(pk))
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You do not have the permission to do that')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -49,8 +108,13 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def delete_room(request, pk):
     room = Room.objects.get(id=int(pk))
+
+    if request.user != room.host:
+            return HttpResponse('You do not have the permission to do that')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
